+++ CHANGELOG.md
@@ -9,7 +9,18 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <br/>
 
 <details>
-<summary><h2> [0.8.1] - 2025/07/15 => 02:35 </h2></summary>
+<summary><h2> [0.8.2] - 2025/07/26 => 19:15 </h2></summary>
+
+- Fix intial child state
+- Fix entries actions
+- <u>Test coverage **_100%_**</u>
+
+</details>
+
+<br/>
+
+<details>
+<summary><h2> [0.8.1] - 2025/07/24 => 02:35 </h2></summary>
 
 - **Enhancement of typings** :
   - better implementation of lib @bemedev/types